flasknn/requirements.txt
flask
flask-socketio
flask_cors
gevent
eventlet
numpy
Pillow
